+++ src/main/java/com/example/backend/domain/alarm/controller/AlarmCommandController.java
@@ -1,7 +1,7 @@
 package com.example.backend.domain.alarm.controller;
 
 import com.example.backend.domain.alarm.service.command.AlarmCommandService;
-import com.example.backend.domain.security.adapter.CustomUserDetails;
+import com.example.backend.domain.auth.jwt.security.CustomUserDetails;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import lombok.RequiredArgsConstructor;
