+++ src/main/java/com/example/backend/domain/alarm/controller/AlarmQueryController.java
@@ -2,7 +2,7 @@ package com.example.backend.domain.alarm.controller;
 
 import com.example.backend.domain.alarm.dto.AlarmResponse;
 import com.example.backend.domain.alarm.service.AlarmQueryService;
-import com.example.backend.domain.security.adapter.CustomUserDetails;
+import com.example.backend.domain.auth.jwt.security.CustomUserDetails;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import lombok.RequiredArgsConstructor;
